requirements.txt
django>=5.0,<5.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
psycopg2-binary>=2.9,<3.0
